add:light control func

+++ RayTracer.h
@@ -12,7 +12,7 @@
 class RayTracer
 {
 private:
-	using PR = function<Color(const double, const double, const Ray&)>;
+	using PR = function<Color(const float, const float, const Ray&)>;
 	Scene *scene;
 	volatile bool isRun = false;
 	thread t[32];
@@ -22,10 +22,10 @@ private:
 
 	void parallelRT(const int8_t tNum, const int8_t tID, const PR &worker);
 	void RTcheck(const int8_t tNum, const int8_t tID);
-	Color RTdepth(const double zNear, const double zFar, const Ray &baseray);
-	Color RTnorm(const double zNear, const double zFar, const Ray &baseray);
-	Color RTtex(const double zNear, const double zFar, const Ray &baseray);
-	Color RTmtl(const double zNear, const double zFar, const Ray &baseray);
+	Color RTdepth(const float zNear, const float zFar, const Ray &baseray);
+	Color RTnorm(const float zNear, const float zFar, const Ray &baseray);
+	Color RTtex(const float zNear, const float zFar, const Ray &baseray);
+	Color RTmtl(const float zNear, const float zFar, const Ray &baseray);
 public:
 	GLuint texID;
 	uint8_t *output;
